+++ docs/source/changelog.rst
@@ -21,6 +21,10 @@ This document displays the differences between each release of curious.
 
  - Add :attr:`.Channel.overwrites` for a key-value mapping of overwrites.
 
+ - Add :class:`._Nickname`, and make :attr:`.Member.nickname` an instance of :class:`._Nickname`.
+
+ - Add :meth:`.EventManager.wait_for_manager`.
+
 0.6.0 (Released 2017-11-05)
 ---------------------------